--- damoim/libs/BaseModelViewSet.py
@@ -1,58 +1,0 @@
-from rest_framework.viewsets import ModelViewSet
-from rest_framework.status import HTTP_403_FORBIDDEN
-from damoim.libs.DamoimException import NotAuthenticated, PermissionDenied, AuthenticationFailed
-
-class BaseModelViewSet(ModelViewSet):
-    def __init__(self):
-        super().__init__(self)
-
-    def handle_exception(self, exc):
-        """
-        Handle any exception that occurs, by returning an appropriate response,
-        or re-raising the error.
-        """
-        if isinstance(exc, (NotAuthenticated, AuthenticationFailed)):
-            # WWW-Authenticate header for 401 responses, else coerce to 403
-            auth_header = self.get_authenticate_header(self.request)
-
-            if auth_header:
-                exc.auth_header = auth_header
-            else:
-                exc.status_code = HTTP_403_FORBIDDEN
-
-        exception_handler = self.get_exception_handler()
-
-        context = self.get_exception_handler_context()
-        response = exception_handler(exc, context)
-
-        if response is None:
-            self.raise_uncaught_exception(exc)
-
-        response.exception = True
-        return response
-
-    def permission_denied(self, request, message=None, code=None):
-        if request.authenticators and not request.successful_authenticator:
-            raise NotAuthenticated()
-        raise PermissionDenied(detail=message, code=code)
-
-    def get_exception_handler(self):
-        return self.settings.EXCEPTION_HANDLER
-
-    def check_object_permissions(self, request, obj):
-        for permission in self.get_permissions():
-            if not permission.has_object_permission(request, self, obj):
-                self.permission_denied(
-                    request,
-                    message=getattr(permission, 'message', None),
-                    code=getattr(permission, 'code', None)
-                )
-
-    def check_permissions(self, request):
-        for permission in self.get_permissions():
-            if not permission.has_permission(request, self):
-                self.permission_denied(
-                    request,
-                    message=getattr(permission, 'message', None),
-                    code=getattr(permission, 'code', None)
-                )
