+++ damoim_project/libs/STATIC_LITERAL_VALUE.py
