+++ damoim_project/user/views.py
@@ -5,11 +5,11 @@ from rest_framework_simplejwt.tokens import RefreshToken
 
 from django.contrib.auth import get_user_model
 
-from damoim.libs.BaseModelViewSet import BaseModelViewSet
-from damoim.libs.Logger import Logging
-from damoim.libs.BaseTokenView import BaseTokenObtainView
-from damoim.user.domain.model.serializer import UserSerializer, CreateUserSerializer
-from damoim.libs.ReturnResponse import ReturnResponse
+from libs.base.views import BaseModelViewSet, BaseTokenObtainView
+from libs.Logger import Logging
+from user.domain.model.serializer import UserSerializer, CreateUserSerializer
+from libs.base.response import ReturnResponse
+from libs.base.exceptions import NotAuthenticated, ClientValueError, AuthenticationFailed, UnknownError
 
 #일단 로그인, 토큰 갱신까지 만들고 Users 데이터 베이스로 사용자 프로필 데이터베이스 작성해야함
 # Todo 유저 관련 데이터 생각하고 다이어그램 그려서 DB 작성하기
@@ -30,18 +30,19 @@ class Login(BaseTokenObtainView):
                 )
                 serialized_user = self.user_serializer_class(user)
                 response.data.update(serialized_user.data)
+            else:
+                raise AuthenticationFailed
             return response
         except Exception as e:
-            logger.exception(e, "Login Failure")
-            return 
-
+            raise ClientValueError
 class Register(BaseModelViewSet):
+    permission_classes = [AllowAny]
     def post(self, request):
         try:
             username = request.data.get("username")
             password = request.data.get("password")
         except Exception as e:
-            logger.exception(e, "error")
+            raise ClientValueError
         serializer = CreateUserSerializer(
             data={
                 "username": username,
@@ -51,6 +52,8 @@ class Register(BaseModelViewSet):
         if serializer.is_valid():
             serializer.save()
             return ReturnResponse(code="0000", flag=True, data=serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            raise ClientValueError
 
 
 
@@ -62,17 +65,16 @@ class LogoutView(BaseModelViewSet):
             token = RefreshToken(refresh_token)
             token.blacklist()
 
-            return Response(status=status.HTTP_205_RESET_CONTENT)
+            return ReturnResponse(code="0000",flag=True,status=status.HTTP_205_RESET_CONTENT)
         except Exception as e:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            raise UnknownError
 
 class RefreshAPI(BaseTokenObtainView):
     permission_classes = [IsAuthenticated, ]
-
     def post(self, request, *args, **kwargs):
         try:
             response = super().post(request, *args, **kwargs)
             return response
         except Exception as e:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
+            return NotAuthenticated
