+++ requirements.txt
@@ -1,0 +1,25 @@
+asgiref==3.6.0
+attrs==22.2.0
+black==23.1.0
+click==8.1.3
+Django==4.1.6
+djangorestframework==3.14.0
+djangorestframework-simplejwt==5.2.2
+django-dotenv==1.4.2
+drf-spectacular==0.25.1
+inflection==0.5.1
+jsonschema==4.17.3
+mypy-extensions==1.0.0
+mysqlclient==2.1.1
+packaging==23.0
+pathspec==0.11.0
+platformdirs==3.0.0
+PyJWT==2.6.0
+PyMySQL==1.0.2
+pyrsistent==0.19.3
+python-dotenv==0.21.1
+pytz==2022.7.1
+PyYAML==6.0
+sqlparse==0.4.3
+tomli==2.0.1
+uritemplate==4.1.1
